Remove warning.
git-svn-id: http://subversion.jackaudio.org/jack/jack2/trunk/jackmp@4011 0c269be4-1314-0410-8aa9-9f06e86f4224

+++ example-clients/ipload.c
@@ -158,7 +158,7 @@ main (int argc, char *argv[])
 					       &status, load_name, load_init);
 	if (status & JackFailure) {
 		fprintf (stderr, "could not load %s, intclient = %d status = 0x%2.0x\n",
-			 load_name, intclient, status);
+			 load_name, (int)intclient, status);
 		return 2;
 	}
 	if (status & JackNameNotUnique) {
